Samples are classified by CAD and No CAD(Coronary Artery Disease) labels
Imported data set, Preprocessed and upsampled it then Bayesian Network is used to calculate the probability of having the disease at first level. At the second level, ECG features are added and all the records having a high probability of having CAD will be predicted again by XGBoost and is compared with actual values, and the model is evaluated.